gdb: Give a better error when a breakpoint can't be placed in a PIE
One issue that seems to crop up fairly regularly on help sites like
stackoverflow, is a use complaining why something like this doesn't
work:

  $ gdb app.exe
  (gdb) disassemble main
  Dump of assembler code for function main:
     0x00000000000005e5 <+0>:	push   %rbp
     0x00000000000005e6 <+1>:	mov    %rsp,%rbp
     0x00000000000005e9 <+4>:	mov    $0x0,%eax
     0x00000000000005ee <+9>:	pop    %rbp
     0x00000000000005ef <+10>:	retq
  (gdb) break *0x5e5
  Breakpoint 1 at 0x5e5: file app.c, line 20.
  (gdb) run
  Starting program: app.exe
  Warning:
  Cannot insert breakpoint 1.
  Cannot access memory at address 0x5e5

Everything appears to be going fine until the user tries to run GDB,
which can, understandably be a little confusing.

The problem is that the executable was compiled as PIE, and the
dynamic linker has placed main at some address that doesn't match the
addresses (or offsets as they really are now) in the applications
ELF.

The problem is that if the user isn't really that aware of PIC/PIE, or
how it works, then they might not realise they what the problem is.

This commit tries to detect when the user has made this mistake, and
gives a warning message that (hopefully) is more helpful, the user now
sees this:

  (gdb) run
  Starting program: app.exe
  Warning:
  Cannot insert breakpoint 1.
  Cannot access memory at address 0x5e5
  The address is located within a code region of `app.exe' which was dynamically
  relocated at run time.  The code region `0x0000000000000500 ... 0x0000000000000662' is now located
  at `0x0000555555554500 ... 0x0000555555554662', you may need to adjust your breakpoints.

gdb/ChangeLog:

	* breakpoint.c (pic_code_error_check): New function.
	(insert_bp_location): Call new function.

gdb/testsuite/ChangeLog:

	* gdb.base/pie-bp-addr-error.c: New file.
	* gdb.base/pie-bp-addr-error.exp: New file.

Change-Id: Iedea261a54ab0a6ed03ec9b8df06fdbe459f28be