gdb: Track the current frame for each thread
Currently in GDB, each time a user switches between threads, the inner
most frame of the thread being switched to is selected.  In some
situations however, it might be helpful for a user to have GDB
remember which frame was selected in each thread, and restore this
frame as the user switches between threads.

This commit adds the following two commands:

  set restore-selected-frame on|off
  show restore-selected-frame

This new option is off by default, so the default behaviour of GDB is
unchanged.

However, with this option turned on GDB will remember, and restore the
selected frame for each thread.

My initial motivation for this change was to have the thread restored
when switching threads with 'thread <num>', however, as I started to
work on this feature I realised that there were a couple of other
places where the sticky frame would naturally appear.  These are 'info
threads' and 'thread apply all'.

With 'info threads' the output contains a 'Frame' column.  Previously,
this was always the innermost frame, the info threads output was
created by switching to each thread in turn and collecting information
about the thread, this naturally placed us at the innermost frame.
Now, the 'Frame' column displays the _selected_ frame for each
thread.

I struggled to decide if this change was good or not.  In the end I
felt that having 'info threads' display the selected frame would feel
more natural, that's the frame you'll end up in if you switch to that
thread, so it seemed to make sense.  However, it would be easy enough
to force the old behaviour if people would prefer.  Alternatively I
could even investigate adding a switch to 'info threads' that allows
the user to select displaying either the selected frame, or the
innermost frame.

For 'thread apply all', again, we used to always apply to the
innermost frame.  Now it's possible for a user to adjust which frame
will be current when the 'thread apply all' runs - this feels like a
useful change to me.  It's easy enough to quickly restore the
innermost frame if required ('thread apply all -- frame 0') and having
the flexibility to tweak the selected frame in just some threads feels
like a nice advantage.  Again, I could potentially add a command flag
here to force the innermost frame.

gdb/ChangeLog:

	* NEWS: Describe new feature.
	* frame.c (cache_selected_frame_on_thread): New function.
	(select_frame): Call new function.
	* frame.h (get_current_frame): Update comment.
	(get_selected_frame): Update comment.
	* gdbthread.h (class thread_info) <selected_frame_id>: New
	member variable.
	<selected_frame_level>: Likewise.
	(switch_to_thread): Extra parameter.
	* thread.c (switch_to_thread_if_alive): Extra parameter, passed to
	switch_to_thread.
	(scoped_restore_current_thread::restore): Restore the frame either
	from the thread, or from the local object.
	(set_executing_thread): Reset the currently selected frame.
	(restore_selected_frame_per_thread): New file level static variable.
	(show_restore_selected_frame_per_thread): New function.
	(print_thread_info_1): Pass extra parameter to switch_to_thread.
	(switch_to_thread): Take extra parameter, restore the previous
	frame if appropriate.
	(thread_apply_all_command): Pass extra parameter to switch_to_thread.
	(thread_apply_command): Likewise.
	(thread_select): Pass extra parameter to switch_to_thread_if_alive.
	(_initialize_thread): Add new set/show variable.

gdb/doc/ChangeLog:

	* gdb.texinfo (Threads): Add anchor to 'info threads'.  Describe
	the Frame column of 'info threads' more.  Describe which frame is
	selected when switching threads, and document the new option for
	restoring the previously selected frame.

gdb/testsuite/ChangeLog:

	* gdb.threads/restore-selected-frame.c: New file.
	* gdb.threads/restore-selected-frame.exp: New file.