gdb: Track the current frame for each thread
Currently in GDB the selected frame is only tracked while a user is
within a selected thread.  When the user switches thread, the selected
frame is reset.

After this commit GDB records the selected frame within its
thread_info structure, and restores each threads selected frame when
switching thread.

If a thread executes at all then the selected frame is discarded, when
the thread stops then the innermost frame is selected, this matches
the existing behaviour of GDB.

The original motivation behind this commit was so a user could switch
threads using the 'thread <ID>' command, and have their selected frame
be remembered, however, once I started implementing this I realised
that there were a couple of places in GDB where the sticky selected
frame would possibly also appear that I hadn't originally considered.
The first was 'info threads', and the other was 'thread apply all'.

With 'info threads' the output contains a 'Frame' column.  Previously,
this was always the innermost frame, the info threads output was
created by switching to each thread in turn and collecting information
about the thread, this naturally placed us at the innermost frame.
Now, the 'Frame' column displays the _selected_ frame for each
thread.

I struggled to decide if this change was good or not.  In the end I
went with the change, my thinking was that if a user visits a thread,
changes the frame, moves to another thread, and then wants to find the
frame they previously visited, then displaying the selected frame
might make this easier.  However, I'm still on the fence for if this
is a good change or not, so would welcome opinions on this.

For 'thread apply all', again, we used to always apply to the
innermost frame.  Now it's possible for a user to adjust which frame
will be current when the 'thread apply all' runs - this feels like a
useful change to me.  Again, I'm happy to discuss this further if
people dislike this change.  Further, 'thread apply all' will now
allow the current frame to be adjusted within it's applied command,
and this change of frame is sticky.

There is, of course, a new flag to turn this new behaviour off, the
new commands are:

  set restore-selected-frame on|off
  show restore-selected-frame

I've turned this new feature on by default for now, but again, I'm
open to being convinced that it should be off by default.

gdb/ChangeLog:

	* NEWS: Describe new feature.
	* frame.c (cache_selected_frame_on_thread): New function.
	(select_frame): Call new function.
	* gdbthread.h (class thread_info) <selected_frame_info>: New
	member variable.
	(switch_to_thread): Extra parameter.
	* thread.c (switch_to_thread_if_alive): Extra parameter, passed to
	switch_to_thread.
	(scoped_restore_current_thread::restore): Restore the frame either
	from the thread, or from the local object.
	(set_executing_thread): Reset the currently selected frame.
	(restore_selected_frame_per_thread): New file level static variable.
	(show_restore_selected_frame_per_thread): New function.
	(print_thread_info_1): Pass extra parameter to switch_to_thread.
	(switch_to_thread): Take extra parameter, restore the previous
	frame if appropriate.
	(thread_apply_all_command): Pass extra parameter to switch_to_thread.
	(thread_apply_command): Likewise.
	(thread_select): Pass extra parameter to switch_to_thread_if_alive.
	(_initialize_thread): Add new set/show variable.

gdb/doc/ChangeLog:

	* gdb.texinfo (Threads): Extend description of 'Frame' field in
	info threads output.  Add description of how GDB tracks the
	previous frame.
	(Selection): Add description of how GDB tracks the selected frame
	per thread, and describe the new commands to disable this
	tracking.

gdb/testsuite/ChangeLog:

	* gdb.mi/user-selected-context-sync.exp: Update expected results.
	* gdb.threads/restore-selected-frame.c: New file.
	* gdb.threads/restore-selected-frame.exp: New file.