Add git sha information to the gdb version string
Adds information about the precise git version from which gdb was
built.  The new version string will look like this:

  GNU gdb (GDB) 10.0.50.20200307-git (137acc6fbd6-72fbdf834da-2-dirty)

The first half (up to the '-git' string) is as before, and taken from
the version.in file within the gdb source tree, with the date text
taken from the bfd source tree.

The second part is new, and is built from the current git version.
This is broken down like this:

  137acc6fbd6-72fbdf834da-2-dirty
  |         | |         | | |   |
  '---A-----' '-----B---' C '-D-'

Where:
 A - This is the last git commit sha.  This represents the HEAD of the
     branch that was built.
 B - This is the merge base between the current branch and the
     upstream master branch.
 C - The number of commits from the merge base to the current sha.
     This gives an impression of how diverged the branch is.
 D - Is the current git tree fully committed? If not then it is dirty.

Parts B, C, and D are optional, though B and C will almost always
appear together (see below for details).  The following are all valid:

  137acc6fbd6-72fbdf834da-2-dirty

Current HEAD is 137acc6fbd6, which is 2 commits from the merge-base
72fbdf834da (which is in sourceware's master branch), however, the
repository that built this GDB was not fully committed, so 137acc6fbd6
will not accurately represent the source code that built this GDB.

  137acc6fbd6-72fbdf834da-2

Like the above, but the repository was fully committed, so 137acc6fbd6
does exactly match the source code that built this version of GDB.

  72fbdf834da-dirty

This version of GDB was built directly from sourceware's master
branch (commit 72fbdf834da), however the repository had local changes
at the time of build, so 72fbdf834da does not exactly match the source
code that built this version of GDB.

  72fbdf834da

Like the above, but the repository was clean at the time of build, so
72fbdf834da exactly matches that source code that built this version
of GDB.

  137acc6fbd6-unknown-dirty
  137acc6fbd6-unknown

These occur when the version script can't find the remote sourceware
repository, and so is unable to figure out a suitable merge-base, nor
can the script figure out if the current HEAD is in sourceware's
master branch.  In this case parts B and C are replaced with the
string 'unknown'.

I have tried to ensure that if the source code is not a git repository
then the git version token will not be added, so folks building from
tar files should get exactly what they had before.

The file gdbsupport/remote-repository contains the pattern used to
identify the remote repository, and the name of the upstream branch
from that repository which we care about.  For us this will be
sourceware and master, but by moving these strings into a separate
file anyone maintaining an out of tree GDB can easily update these to
point to their repository and branch, and get project specific version
strings.

gdbsupport/ChangeLog:

	* create-version.sh: Add git sha information.
	* remote-repository: New file.

gdbserver/ChangeLog:

	* Makefile.in (version-generated.cc): Add remote-repository file
	as a dependency.

gdb/ChangeLog:

	* Makefile.in (stamp-version): Add remote-repository file as a
	dependency.