gdb: Restore previously selected thread when switching inferior
This commit changes how GDB behaves when switching between
multi-threaded inferiors.

Previously when switching GDB would select "any" thread from the
inferior being switched to.  In practice this usually means the thread
with the lowest thread number.

After this commit GDB remembers which thread the user had selected
within the inferior and tries to reselect that thread when switching
back to an inferior.  The assumption is that the selected thread is
the one that the user is most interested in, and so switching back to
that makes more sense than selecting some other thread.

If the previously selected thread is no longer available, for example,
if the thread has exited, then GDB will fall back on the old
behaviour.

I did consider making this behaviour switchable, but the new behaviour
seems (to me) obviously better, so I haven't added a switch for it.
However, if people feel the old behaviour is worth preserving then I'm
happy to add a switch.

The other thing I considered, but didn't do is to add a warning when
switching inferiors if the previously selected thread is no longer
available.  My reasoning here is that GDB should already have informed
the user that the thread has exited, and there is already a message
indicating which thread has been switched too, so adding an extra
warning felt like unneeded clutter, but again, if people feel strongly
that there should be a warning, I'm happy to add one.

One test needed to be updated after this change, and there's a new
test that covers this behaviour in more depth.

gdb/ChangeLog:

	* inferior.c (inferior_command): Store current inferior_ptid
	before switching inferiors.  Reselect the previous inferior_ptid
	if possible after switching to the new inferior.
	* inferior.h (class inferior) <previous_inferior_ptid>: New member
	variable.
	* NEWS: Mention new feature.

gdb/testsuite/ChangeLog:

	* gdb.mi/user-selected-context-sync.exp: Update expected results.
	* gdb.threads/restore-thread.c: New file.
	* gdb.threads/restore-thread.exp: New file.

gdb/doc/ChangeLog:

	* gdb.texinfo (Inferiors Connections and Programs): Mention thread
	tracking within the inferior command.
	(Threads): Mention thread tracking in the general thread
	discussion.