gdb: Restore previously selected thread when switching inferior
This commit adds a new option that allows the user to control how GDB
behaves when switching between multi-threaded inferiors.

Currently (and this remains the default after this commit) when
switching between inferiors GDB would select the first non-exited
thread from the inferior being switched to.

This commit adds the following new commands:

     set restore-selected-thread on|off
     show restore-selected-thread

This option is off by default in order to retain the existing
behaviour, but, when switched on GDB will remember which thread was
selected in each inferior.  As the user switches between inferiors GDB
will attempt to restore the previously selected thread.

If the previously selected thread is no longer available, for example,
if the thread has exited, then GDB will fall back on the old
behaviour.

I did consider, but eventually didn't implemented, adding a warning
when switching inferiors if the previously selected thread is no
longer available.  My reasoning here is that GDB should already have
informed the user that the thread has exited, and there is already a
message indicating which thread has been switched too, so adding an
extra warning felt like unneeded clutter.

In order to store the thread within the inferior I store a pointer to
the thread_info object of the previously selected thread.  When
fetching the thread_info it is important that we do actually have a
current thread otherwise this happens:

  $ gdb
  (gdb) add-inferior
  (gdb) inferior 2
  ./gdb/thread.c:95: internal-error: thread_info* inferior_thread(): Assertion `current_thread_ != nullptr' failed.

To avoid this I added a check that inferior_ptid is not null_ptid.
Though it is not always the case, there are plenty of places in GDB
where a call to inferior_thread () is guarded by such a check.

There's a new test for this functionality.

gdb/ChangeLog:

	* inferior.c (inferior_command): Store current thread_info before
	switching inferiors.  Reselect the previous thread_info if
	possible after switching to the new inferior.
	(initialize_inferiors): Register restore-selected-thread option.
	* inferior.h (class inferior) <previous_thread_info>: New member
	variable.
	* NEWS: Mention new feature.

gdb/testsuite/ChangeLog:

	* gdb.threads/restore-thread.c: New file.
	* gdb.threads/restore-thread.exp: New file.

gdb/doc/ChangeLog:

	* gdb.texinfo (Inferiors Connections and Programs): Mention thread
	tracking within the inferior command.
	(Threads): Mention thread tracking in the general thread
	discussion.