gdb: ensure location types match before swapping locations
In the following conditions:

  - A target with hardware breakpoints available, and
  - A target that uses software single stepping,
  - An instruction at ADDRESS loops back to itself,

Now consider the following steps:

  1. The user places a hardware breakpoint at ADDRESS (an instruction
  that loops to itself),

  2. The inferior runs and hits the breakpoint at ADDRESS,

  3. The user tells GDB to 'continue'.

In #3 when the user tells GDB to continue, GDB first disables the
hardware breakpoint at ADDRESS, and then inserts a software single
step breakpoint at ADDRESS.  The original user created breakpoint was
a hardware breakpoint, while the single step breakpoint will be a
software breakpoint.

GDB continues and immediately hits the software single step
breakpoint.

GDB then deletes the software single step breakpoint by calling
delete_single_step_breakpoints, which eventually calls
delete_breakpoint, which, once the breakpoint (and its locations) are
deleted, calls update_global_location_list.

During update_global_location_list GDB spots that we have an old
location (the software single step breakpoint location) that is
inserted, but being deleted, and a location (the original hardware
breakpoint) at the same address which we are keeping, but which is not
currently inserted, GDB then calls breakpoint_locations_match on these
two locations.

Currently the locations do match, and so GDB calls swap_insertion
which swaps the "inserted" state of the two locations.  The user
created hardware breakpoint is marked as inserted, while the GDB
internal software single step breakpoint is now marked as not
inserted.  After this GDB returns through the call stack and leaves
delete_single_step_breakpoints.

After this GDB continues with its normal "stopping" process, as part
of this stopping process GDB removes all the breakpoints from the
target.  Due to the swap it is now the user created hardware
breakpoint that is marked as inserted, so it is this breakpoint GDB
tries to remove.

The problem is that GDB inserted the software single step breakpoint
as a software breakpoint, but is now trying to remove the hardware
breakpoint.  The problem is removing a software breakpoint is very
different to removing a hardware breakpoint, this could result is some
undetected undefined behaviour, or as in the original bug report (PR
gdb/25741), could result in the target throwing an error.

This original patch proposed by Keno Fischer is here:

  https://sourceware.org/pipermail/gdb-patches/2020-April/167202.html

This patch modified breakpoint_locations_match so that two breakpoints
of different types would no longer match.  This resolves the immediate
problem described above, but introduces some other issues.

The breakpoint_locations_match function is also used to control
whether two breakpoints places at the same address should both be
inserted or not.

Consider this (slightly contrived) use case:

  (gdb) set debug remote 1
  (gdb) set breakpoint  always-inserted on
  (gdb) hbreak  MyFunc0
  Sending packet: $me0000202,2#84...Packet received: 0000
  Hardware assisted breakpoint 3 at 0xe0000202: file test.c, line 288.
  Sending packet: $m80000188,1#63...Packet received: 00
  Sending packet: $Z1,e0000202,2#ce...Packet received: OK
  (gdb) break MyFunc0
  Note: breakpoint 3 also set at pc 0xe0000202.
  Sending packet: $me0000202,2#84...Packet received: 0000
  Breakpoint 4 at 0xe0000202: file test.c, line 288.
  Sending packet: $m80000188,1#63...Packet received: 00

Notice after setting the hardware breakpoint we see the Z1 packet
sent, but after setting the software breakpoint there's no Z0 packet,
GDB sees that the locations match and doesn't make the second
insertion.  If we modify breakpoint_locations_match then we would
insert both a hardware _and_ software breakpoint.  This probably
doesn't matter, but is not ideal, so a solution that doesn't do this
would be better.

The proposal here is to focus specifically on the case where we are
considering swapping the inserted status of two breakpoint locations,
and moves the location type check out of breakpoint_locations_match,
and into the caller.  This resolves the original issue, while avoiding
the double insertion problem.

gdb/ChangeLog:

	PR gdb/25741
	* breakpoint.c (breakpoint_locations_match): Compare location
	types.