gdb/testsuite: Check for syntax errors in some python tests
While writing a Python test making use of gdb_py_test_silent_cmd I
made a typo and introduced a syntax error into my test.  When I ran
the test script though, the test still passed!

The reason is that gdb_py_test_silent_cmd looks for the pattern
'Traceback' indicating an exception was thrown, or the prompt,
indicating the command completed.

This commit extends this command, drawing for inspiration on
'gdb_test_no_output' (from lib/gdb.exp).  Now, we check that there is
no output from the python command; previously any output would be
ignored, despite this function declaring that this was a silent
command.

I also added an extra case to check specifically for 'SyntaxError' in
the output, and mark this as a failure.

gdb/testsuite/ChangeLog:

	* lib/gdb-python.exp (gdb_py_test_silent_cmd): A pass should
	produce no output. A SyntaxError is a specific failure case.