gdb: Print better error messages for maint print * register commands
The regcache_print function is the worker core for many maintenance
commands.  This commit makes its error message command specific.

I removed internationalisation of the string used in the error message
as this string is a command name.  To me it makes no sense to
internationalise the command name in an error message when the use has
just entered the non-internationalised command name in order to
trigger the error.

gdb/ChangeLog:

	* regcache-dump.c (regcache_print_open_perror): New function.
	(regcache_print): Use new function.