gdb/testsuite: Add missing expected results
The tests in this script are driven from two lists of expected
results, one of the lists is missing some data so DejaGNU ends up
matching against the empty string (which passes).

This commit adds the missing expected results into the script.

I could rewrite this test to make things more robust, however, a later
commit is going to completely rewrite this test script, I'm simply
adding this here so that _before_ the rewrite the test is complete,
then if anyone ever digs into the history of this test script things
will make sense (I hope).

gdb/testsuite/ChangeLog:

	* gdb.fortran/array-slices.exp: Add missing message data.