gdb: Don't let SIGWINCH interrupt waiting for remote target
When connecting to a remote target a SIGWINCH will cause GDB to stop
waiting an complain about an interrupted system call, so:

  (gdb) target remote :1234
  # User resizes termainal...
  :1234: Interrupted system call.
  (gdb)

While investigating I also discovered that hitting Ctrl-C looks like
this:

  (gdb) target remote :1234
  # User hits Ctrl-C
  ^C:1234: Interrupted system call.
  (gdb) Quit
  (gdb)

Which doesn't seem great.

In this commit I wrap the code that waits for a connection in a loop
that will swallow any EINTR errors.  As a result SIGWINCH no longer
causes the connection loop to break, but nor does Ctrl-C.  So I add a
check of the QUIT flag, and now we get:

  (gdb) target remote :1234
  # User resizes termainal, and GDB continues to wait.

And also:

  (gdb) target remote :1234
  # User hits Ctrl-C
  ^CQuit
  (gdb)

Which seems better.

There is a test for this, which I've tried to make resilient against
timing issues, but I don't know how good it will actually turn out to
be.

gdb/ChangeLog:

	PR remote/26221
	* ser-tcp.c (wait_for_connection): Update comment.  Ignore EINTR
	errors, and handle QUIT from user.

gdb/testsuite/ChangeLog:

	PR remote/26221
	* gdb.server/connect-resize-quit.exp: New file.